# http: simplify checkIsHttpToken()