@Trott 2017-11-30
Benchmark results:

```console
                                                                 improvement confidence      p.value
 http/check_is_http_token.js n=1000000 key=":"                      -67.14 %        *** 4.065668e-34
 http/check_is_http_token.js n=1000000 key=":alternate-protocol"    -54.61 %        *** 5.705110e-28
 http/check_is_http_token.js n=1000000 key="((((())))"              -66.34 %        *** 2.636118e-33
 http/check_is_http_token.js n=1000000 key="@@"                     -67.24 %        *** 2.329542e-28
 http/check_is_http_token.js n=1000000 key="Accept-Ranges"          146.79 %        *** 2.298935e-45
 http/check_is_http_token.js n=1000000 key="alt-svc"                 -3.79 %        *** 3.200379e-04
 http/check_is_http_token.js n=1000000 key="alternate-protocol:"    150.80 %        *** 1.859951e-41
 http/check_is_http_token.js n=1000000 key="alternate-protocol"     239.27 %        *** 2.887460e-42
 http/check_is_http_token.js n=1000000 key="Cache-Control"          148.05 %        *** 1.588821e-41
 http/check_is_http_token.js n=1000000 key="Connection"              38.53 %        *** 9.847708e-29
 http/check_is_http_token.js n=1000000 key="Content-Encoding"       205.91 %        *** 1.416987e-37
 http/check_is_http_token.js n=1000000 key="content-length"         157.29 %        *** 1.049744e-37
 http/check_is_http_token.js n=1000000 key="Content-Location"       203.87 %        *** 1.005885e-38
 http/check_is_http_token.js n=1000000 key="content-type"            60.62 %        *** 1.051226e-40
 http/check_is_http_token.js n=1000000 key="Content-Type"            61.86 %        *** 1.993077e-30
 http/check_is_http_token.js n=1000000 key="date"                   -58.40 %        *** 4.130459e-34
 http/check_is_http_token.js n=1000000 key="ETag"                   -58.21 %        *** 2.484249e-43
 http/check_is_http_token.js n=1000000 key="Expires"                 -1.77 %            8.884764e-02
 http/check_is_http_token.js n=1000000 key="Keep-Alive"              38.08 %        *** 3.190244e-35
 http/check_is_http_token.js n=1000000 key="Last-Modified"          150.14 %        *** 4.677389e-55
 http/check_is_http_token.js n=1000000 key="location"                12.19 %        *** 1.952609e-14
 http/check_is_http_token.js n=1000000 key="server"                 -18.56 %        *** 6.947429e-20
 http/check_is_http_token.js n=1000000 key="Server"                 -18.91 %        *** 1.729007e-19
 http/check_is_http_token.js n=1000000 key="status"                 -19.17 %        *** 4.218003e-27
 http/check_is_http_token.js n=1000000 key="TCN"                    -61.80 %        *** 1.219346e-23
 http/check_is_http_token.js n=1000000 key="Transfer-Encoding"      223.55 %        *** 1.359534e-40
 http/check_is_http_token.js n=1000000 key="Vary"                   -59.13 %        *** 3.830565e-45
 http/check_is_http_token.js n=1000000 key="version"                 -6.03 %        *** 2.214569e-04
 http/check_is_http_token.js n=1000000 key="x-frame-options"        164.26 %        *** 1.243066e-30
 http/check_is_http_token.js n=1000000 key="x-xss-protection"       204.88 %        *** 1.000083e-35
 http/check_is_http_token.js n=1000000 key="中文呢"                 -31.09 %        *** 6.800282e-23
```