@Trott 2017-11-30
Benchmark summary seems to be:

* Longer valid tokens perform much better with this change.
* Shorter valid tokens take a perf hit with this change.
* Invalid tokens where the invalid character appear early on in the string perform better with the existing code, but you get the opposite result if it appears late in the string.

I'm tempted to log arguments sent to `checkIsHttpToken()` in a real world app to see what kinds of strings are processed and how often.