@mscdex 2017-11-30
I think I'd prefer a hybrid + expanded loop unrolling solution. The expanded unrolling would cover the most common http headers (based on headers listed on Wikipedia for example and also the list we use in _http_incoming.js when converting to headers to lowercase) and the regexp would be used for larger header names.

We could dynamically generate the function to avoid the lengthy function, it performs the same as the inline version:

<details>

```js
const validTokens = [
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, // 0 - 15
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, // 16 - 31
  0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, // 32 - 47
  1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, // 48 - 63
  0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, // 64 - 79
  1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, // 80 - 95
  1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, // 96 - 111
  1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, // 112 - 127
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, // 128 ...
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0  // ... 255
];
const tokenRegExp = /^[\^_`a-zA-Z\-0-9!#$%&'*+.|~]+$/;
function checkIsHttpToken(val) {
  if (val.length > 19) return tokenRegExp.test(val);
  if (!validTokens[val.charCodeAt(0)]) return false;
  if (val.length < 2) return true;
  if (!validTokens[val.charCodeAt(1)]) return false;
  if (val.length < 3) return true;
  if (!validTokens[val.charCodeAt(2)]) return false;
  if (val.length < 4) return true;
  if (!validTokens[val.charCodeAt(3)]) return false;
  if (val.length < 5) return true;
  if (!validTokens[val.charCodeAt(4)]) return false;
  if (val.length < 6) return true;
  if (!validTokens[val.charCodeAt(5)]) return false;
  if (val.length < 7) return true;
  if (!validTokens[val.charCodeAt(6)]) return false;
  if (val.length < 8) return true;
  if (!validTokens[val.charCodeAt(7)]) return false;
  if (val.length < 9) return true;
  if (!validTokens[val.charCodeAt(8)]) return false;
  if (val.length < 10) return true;
  if (!validTokens[val.charCodeAt(9)]) return false;
  if (val.length < 11) return true;
  if (!validTokens[val.charCodeAt(10)]) return false;
  if (val.length < 12) return true;
  if (!validTokens[val.charCodeAt(11)]) return false;
  if (val.length < 13) return true;
  if (!validTokens[val.charCodeAt(12)]) return false;
  if (val.length < 14) return true;
  if (!validTokens[val.charCodeAt(13)]) return false;
  if (val.length < 15) return true;
  if (!validTokens[val.charCodeAt(14)]) return false;
  if (val.length < 16) return true;
  if (!validTokens[val.charCodeAt(15)]) return false;
  if (val.length < 17) return true;
  if (!validTokens[val.charCodeAt(16)]) return false;
  if (val.length < 18) return true;
  if (!validTokens[val.charCodeAt(17)]) return false;
  if (val.length < 19) return true;
  if (!validTokens[val.charCodeAt(18)]) return false;
  return true;
}
```
</details>

Benchmarking the various solutions with `1e8` iterations, I see these results for various headers:

|    | master | this PR | hybrid+expanded unroll |
| - | - | - | - |
| If-Unmodified-Since (19 chars) | 10.7s | 5.6s | 3.4s |
| Content-Type (12 chars) | 6.1s | 4.3s | 2.4s |
| Date (4 chars) | 0.96s | 3.0s | 1.1s |